add FieldType entity
