+++ src/forms/forms.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { FormsService } from './forms.service';
 import { FormsController } from './forms.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FormRepository } from './form.repository';
+import { FormRepository } from './repositories/form.repository';
 
 @Module({
     imports: [TypeOrmModule.forFeature([FormRepository])],
